ISSUE #6034 - Refactored kanbanboard to be a bit more readable
