+++ frontend/src/v5/ui/components/kanbanBoard/laneColumn.tsx
@@ -16,92 +16,10 @@
  */
 
 import { Fragment, memo, useCallback, useEffect, useLayoutEffect, useMemo, useState } from 'react';
-import { Box, Stack, Typography } from '@mui/material';
-import { alpha, styled } from '@mui/material/styles';
-import { getRadius } from './KanbanBoardComponents';
-import type { DropTargetStore, LaneColumnProps } from './KanbanBoardTypes';
 
-const LanePaper = styled('div', {
-	shouldForwardProp: (prop) => prop !== 'isDropTarget',
-})<{ isDropTarget: boolean }>(({ isDropTarget, theme }) => ({
-	width: '400px',
-	flexShrink: 0,
-	display: 'block',
-	flexDirection: 'column',
-	borderRadius: getRadius(theme, 2),
-	border: `1px solid ${
-		isDropTarget ? theme.palette.primary.main : (theme.palette as any).base.lightest
-	}`,
-	backgroundColor: isDropTarget
-		? alpha(theme.palette.primary.main, 0.08)
-		: theme.palette.background.paper,
-	transition: 'background-color 160ms ease, border-color 160ms ease',
-	overflowY: 'hidden',
-	maxHeight: '100%',
-	padding: 8,
-}));
+import type { DropTargetStore, LaneColumnProps } from './kanbanBoard.types';
+import { CardShell, DropPlaceholder, LaneCards, LaneHeader, LanePaper, LaneSubtitle, LaneTitle, LaneTitleGroup } from './kanbanBoard.styles';
 
-const LaneHeader = styled(Stack)(({ theme }) => ({
-	flexDirection: 'row',
-	justifyContent: 'space-between',
-	alignItems: 'center',
-	gap: theme.spacing(1),
-	padding: theme.spacing(1.25, 1.5),
-}));
-
-const LaneTitleGroup = styled(Box)({
-	minWidth: 0,
-});
-
-const LaneTitle = styled(Typography)({
-	fontWeight: 800,
-	overflowWrap: 'anywhere',
-});
-
-const LaneSubtitle = styled(Typography)({
-	display: 'block',
-	overflowWrap: 'anywhere',
-});
-
-const LaneCards = styled(Stack)(() => ({
-	flex: 1,
-	overflowY: 'auto',
-	gap: 6,
-	maxHeight: '100%',
-}));
-
-const DropPlaceholder = styled(Box, {
-	shouldForwardProp: (prop) => prop !== 'placeholderHeight',
-})<{ placeholderHeight: number }>(({ placeholderHeight, theme }) => ({
-	height: placeholderHeight,
-	flexShrink: 0,
-	border: '2px dashed',
-	borderColor: alpha(theme.palette.primary.main, 0.48),
-	borderRadius: getRadius(theme, 2),
-	backgroundColor: alpha(theme.palette.primary.main, 0.08),
-	boxShadow: `inset 0 0 0 1px ${alpha(theme.palette.primary.main, 0.12)}`,
-	transition:
-		'background-color 160ms ease, border-color 160ms ease, height 160ms ease',
-}));
-
-const CardShell = styled(Box, {
-	shouldForwardProp: (prop) =>
-		prop !== 'isDraggable' && prop !== 'canClickCards',
-})<{ isDraggable: boolean; canClickCards: boolean }>(
-	({ canClickCards, isDraggable, theme }) => ({
-		cursor: isDraggable ? 'grab' : canClickCards ? 'pointer' : 'default',
-		opacity: isDraggable ? 1 : 0.72,
-		outline: 0,
-		touchAction: isDraggable ? 'none' : 'auto',
-		'&:active': {
-			cursor: isDraggable ? 'grabbing' : canClickCards ? 'pointer' : 'default',
-		},
-		'&:focus-visible': {
-			borderRadius: getRadius(theme, 2),
-			boxShadow: `0 0 0 3px ${alpha(theme.palette.primary.main, 0.32)}`,
-		},
-	}),
-);
 
 function useLaneDropState(store: DropTargetStore, laneId: string) {
 	const [laneDropState, setLaneDropState] = useState(() =>
